Add rudimentary argfile support to `zinc`
Mostly cribbed from the `scalac` implementation, this adds support
for `@filename` arguments which point to files which contain
additional arguments, one per line, which should be treated as
though they were specified on the command line.  This will be
use to circumvent kernel limits on command line size for
classpaths, for instance.