Add fixed-output derivation for patched `zinc`.
This is mostly modeled after the `sbtix` build command, excluding
the stuff there which handles making the dependency stuff
deterministic and hermetic (by putting the deps themselves into
the store).

Also, this derivation makes an unpatched copy of the `zinc` wrapper
script for use on windows, where we'll be wanting to use the
built artifacts from the repo itself, not from a nix store.